+++ psiresp/mixins/__init__.py
@@ -1,5 +1,5 @@
-from .conformer import ConformerGenerator
 from .grid import GridMixin
 from .io import IOMixin
 from .molecule import MoleculeMixin
 from .qm import QMMixin
+from .resp import BaseRespOptions
