+++ psiresp/utils/__init__.py
@@ -1,0 +1,2 @@
+BOHR_TO_ANGSTROM = 0.52917721092
+ANGSTROM_TO_BOHR = 1/BOHR_TO_ANGSTROM
